#7 sctp stack based on com.sun.nio.sctp -les conversions d'endianess specifiques a lksctp sont gérées par la stack lkstcp (dans les classes d'adaptateurs). La convention par defaut des APIs exposées est celle de java. -revoit le code de serialisation pour eviter d'avoir des conversions unsigned short et unsigned int vers String lors de la manipulation des attributs d'un InfoSctp (On peut utiliser les methodes statiques de l'helper InfoSctp.StringPrinter) - Les tests unitaires 'tutorial/sctp/*' passent avec la stack Lkstcp, mais tjrs pas avec la stack com.sun.nio.sctp

+++ branches/t7-sctp/mts/src/main/java/com/devoteam/srit/xmlloader/sctp/sunnio/InfoSunNioSctp.java
@@ -198,7 +198,7 @@ public class InfoSunNioSctp extends InfoSctp {
 	 * @param associationId the AssociationSctp value
 	 */
 	@Override
-	public void setAssociationId( long associationId ) throws Exception{
+	public void setAssociationId( int associationId ) throws Exception{
 		throw new UnsupportedOperationException ("this method is not implemented");
 	}
